WIP: more work on data debit operation logging, recording operations to database and testing

+++ src/main/scala/hatdex/hat/api/models/ApiDataDebit.scala
@@ -37,7 +37,7 @@ case class ApiDataDebitOut(
   endDate: LocalDateTime,
   rolling: Boolean,
   sell: Boolean,
-  price: Double,
+  price: Float,
   kind: String,
   bundleContextless: Option[ApiBundleContextlessData],
   bundleContextual: Option[Seq[ApiEntity]])
